+++ include/filesystem/applicationuserdirectory.h
@@ -17,31 +17,23 @@
  *
  * @section DESCRIPTION
  *
- * Supported user directories.
+ * Supported application user directories.
  */
 
-#ifndef USERDIRECTORY_H
-#define USERDIRECTORY_H
+#ifndef APPLICATIONUSERDIRECTORY_H
+#define APPLICATIONUSERDIRECTORY_H
 
 namespace Nickvision::Filesystem
 {
     /**
-     * @brief Supported user directories.
+     * @brief Supported application user directories.
      */
-    enum class UserDirectory
+    enum class ApplicationUserDirectory
     {
-        Home, ///< The home directory.
-        Config, ///< The configuration directory.
-        Cache, ///< The cache directory.
-        LocalData, ///< The local data directory.
-        Desktop, ///< The desktop directory.
-        Documents, ///< The documents directory.
-        Downloads, ///< The downloads directory.
-        Music, ///< The music directory.
-        Pictures, ///< The pictures directory.
-        Templates, ///< The templates directory.
-        Videos ///< The videos directory.
+        Config, ///< The application configuration directory.
+        Cache, ///< The application cache directory.
+        LocalData ///< The application local data directory.
     };
 }
 
-#endif //USERDIRECTORY_H
+#endif //APPLICATIONUSERDIRECTORY_H