+++ include/filesystem/userdirectories.h
@@ -25,18 +25,26 @@
 
 #include <filesystem>
 #include <string>
+#include "applicationuserdirectory.h"
 #include "userdirectory.h"
 
 namespace Nickvision::Filesystem::UserDirectories
 {
-    /*
+    /**
      * @brief Gets the path of a user directory.
      * @brief Supports XDG directories on Linux.
      * @param directory The UserDirectory to get the path of
-     * @param appName An optional application name to use for application specific directories
      * @return The path of the directory
      */
-    std::filesystem::path get(UserDirectory directory, const std::string& appName = "");
+    std::filesystem::path get(UserDirectory directory);
+    /**
+     * @brief Gets the path of an application user directory.
+     * @brief Supports XDG directories on Linux.
+     * @param directory The ApplicationUserDirectory to get the path of
+     * @param appName The application name
+     * @return The path of the directory
+     */
+    std::filesystem::path get(ApplicationUserDirectory directory, const std::string& appName);
 }
 
 #endif //USERDIRECTORIES_H